requirements.txt
pandapower==2.14.0
streamlit==1.31.0
ollama==0.0.11
networkx==3.2
matplotlib==3.8.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2
joblib==1.3.2
pyyaml==6.0
plotly==5.17.0
prophet==1.1.4
tensorflow==2.13.0
python-dotenv==1.0.0